refactored some BesuCommandTests into separate test classes (#6644)
